requirements.txt
# KOL 评价报告生成器依赖
# 安装: pip install -r requirements.txt

# HTTP 请求库（用于同步请求）
requests>=2.28.0
urllib3>=1.26.0

# 异步 HTTP 请求库（用于异步抓取推文）
aiohttp>=3.8.0

# OpenAI API（用于 AI 评分）
openai>=1.0.0

# 环境变量管理（可选，用于 .env 文件）
python-dotenv>=0.19.0

